requirements.txt
annotated-types==0.6.0
anyio==4.6.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.112.2
fastapi-cli==0.0.7
filelock==3.17.0
fsspec==2025.2.0
google-auth==2.38.0
google-genai==1.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.7
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.37
langchain-ollama==0.2.3
langsmith==0.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.3
ollama==0.4.7
ollama-ocr==0.1.3
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pip==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
safetensors==0.5.2
setuptools==75.8.0
shellingham==1.5.4
sniffio==1.3.0
starlette==0.38.2
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wheel==0.45.1
zstandard==0.23.0
